Updated swiper
