+++ src/modules/Gallery/components/PicturesSwiper/index.js
@@ -1,0 +1,1 @@
+export { default } from './PicturesSwiper';
